simple edition smartQA

+++ src/components/Body/Body.js
@@ -4,6 +4,7 @@ import './Body.css';
 import Item from './Item';
 import axios from 'axios';
 import { intentList } from '../../apiurl'
+import Modal from 'react-modal';
 
 export default class Body extends Component {
 
@@ -16,7 +17,8 @@ export default class Body extends Component {
 
 
     this.state = {
-      itemList: []
+      itemList: [],
+      isPopupOpen: false
     }
     this.addItem = this.addItem.bind(this);
     this.deleteItem = this.deleteItem.bind(this);
@@ -26,7 +28,7 @@ export default class Body extends Component {
     let dataList = [];
     // console.log(this.props.ofelId)
     axios.get(intentList,{headers: {'ofelId': '888'}})
-      .then((res) => res.data.data.map(qaItem => {
+      .then((res) => res.data.data.reverse().map(qaItem => {
         dataList.push(<Item 
                         key={qaItem.id}
                         deleteItem={this.deleteItem}
@@ -44,21 +46,36 @@ export default class Body extends Component {
       })))
       // .then(() => console.log(this.props.ofelId))
     }
+  handlePopupOpen = () => {
+    this.setState({
+      isPopupOpen: true
+    })
+  }
 
+  handlePopupClose = () => {
+    this.setState({
+      isPopupOpen: false
+    })
+  }
   addItem() {
     const items = this.state.itemList ? this.state.itemList : []
     const index = items.findIndex((data) => data.key == items.length); //check Item's key repeat
-    items.push(
-      <Item
-        key={index <0 ? items.length : `${items.length}${index}`}
-        deleteItem={this.deleteItem}
-        keyId={index <0 ? items.length : `${items.length}${index}`}
-        keywords=""
-        fetchApi={this.fetchApi}
-      />
-    )
-    this.setState({itemList: items})
-    
+    if(items.length == this.props.intentQuota) {
+      this.handlePopupOpen();
+    }else{
+      items.unshift(
+        <Item
+          key={index <0 ? items.length : `${items.length}${index}`}
+          deleteItem={this.deleteItem}
+          keyId={index <0 ? items.length : `${items.length}${index}`}
+          keywords=""
+          fetchApi={this.fetchApi}
+          enable={true}
+          forceReveal={true}
+        />
+      )
+      this.setState({itemList: items})
+    }
   }
 
   deleteItem(i) {
@@ -72,7 +89,7 @@ export default class Body extends Component {
     let dataList = [];
     // console.log(this.props.ofelId)
     axios.get(intentList,{headers: {'ofelId': '888'}})
-      .then((res) => res.data.data.map(qaItem => {
+      .then((res) => res.data.data.reverse().map(qaItem => {
         dataList.push(<Item 
                         key={qaItem.id}
                         deleteItem={this.deleteItem}
@@ -95,7 +112,13 @@ export default class Body extends Component {
     
     return (
       <div className="body">
-        <New_Button count={this.state.itemList ? this.state.itemList.length : 0} addItem={this.addItem} intentQuota={this.props.intentQuota}/>
+        <New_Button 
+          count={this.state.itemList ? this.state.itemList.length : 0} 
+          addItem={this.addItem} 
+          intentQuota={this.props.intentQuota} 
+          isPopupOpen={this.state.isPopupOpen} 
+          handlePopupOpen={this.handlePopupOpen} 
+          handlePopupClose={this.handlePopupClose}/>
         <div style={{width: '100%'}}>
           {this.state.itemList}
         </div>
